pyproject.toml
[tool.poetry]
name = "informatica"
version = "0.1.0"
description = "Conversate with documents"
authors = ["Lucian Musat <amibadesign@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.39.0"
pypdf2 = "^3.0.1"
langchain = "^0.3.2"
langchain-weaviate = "^0.0.3"
langchain-ollama = "^0.2.0"
weaviate-client = "^4.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
